+++ inc/plugins/class-form-records-export.php
@@ -1,0 +1,54 @@
+<?php
+/**
+ * Form Submission Record export.
+ *
+ * Handles the admin-ajax bulk export of Submission Records. Exporting is an Otter Pro
+ * feature.
+ *
+ * @package ThemeIsle\GutenbergBlocks\Plugins
+ */
+
+namespace ThemeIsle\GutenbergBlocks\Plugins;
+
+use ThemeIsle\GutenbergBlocks\Pro;
+
+/**
+ * Class Form_Records_Export
+ */
+class Form_Records_Export {
+	/**
+	 * Register hooks.
+	 *
+	 * @return void
+	 */
+	public function register() {
+		add_action( 'wp_ajax_otter_form_submissions', array( $this, 'export_submissions' ) );
+	}
+
+	/**
+	 * Export submissions with ajax. Bulk export is a Pro feature.
+	 */
+	public function export_submissions() {
+		if ( ! Pro::is_pro_active() ) {
+			wp_die( esc_html( __( 'Exporting submissions requires Otter Pro.', 'otter-blocks' ) ) );
+		}
+
+		$nonce = isset( $_POST['_nonce'] ) ? sanitize_text_field( $_POST['_nonce'] ) : ''; // phpcs:ignore WordPress.Security.NonceVerification.Recommended
+		if ( ! wp_verify_nonce( $nonce, 'otter_form_export_submissions' ) ) {
+			wp_die( esc_html( __( 'Invalid nonce.', 'otter-blocks' ) ) );
+		}
+
+		if ( ! current_user_can( 'manage_options' ) ) {
+			wp_die( esc_html( __( 'You are not allowed to export submissions.', 'otter-blocks' ) ) );
+		}
+
+		// Export submissions.
+		require_once ABSPATH . 'wp-admin/includes/export.php';
+		ob_start();
+		export_wp( array( 'content' => Form_Submissions::FORM_RECORD_TYPE ) );
+		$export = ob_get_clean();
+
+		echo ent2ncr( $export ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
+		wp_die();
+	}
+}
