+++ inc/plugins/class-form-submissions.php
@@ -14,7 +14,6 @@ namespace ThemeIsle\GutenbergBlocks\Plugins;
 use ThemeIsle\GutenbergBlocks\Integration\Form_Data_Request;
 use ThemeIsle\GutenbergBlocks\Integration\Form_Data_Response;
 use ThemeIsle\GutenbergBlocks\Integration\Form_Settings_Data;
-use ThemeIsle\GutenbergBlocks\Pro;
 use ThemeIsle\GutenbergBlocks\Server\Form_Server;
 use WP_Post;
 use WP_Query;
@@ -123,7 +122,7 @@ class Form_Submissions {
 		add_action( 'otter_form_update_record_meta_dump', array( $this, 'update_submission_dump_data' ), 10, 2 );
 		add_action( 'otter_form_automatic_confirmation', array( $this, 'move_old_stripe_draft_sessions_to_unread' ) );
 
-		add_action( 'wp_ajax_otter_form_submissions', array( $this, 'export_submissions' ) );
+		( new Form_Records_Export() )->register();
 	}
 
 	/**
@@ -815,33 +814,6 @@ class Form_Submissions {
 		);
 
 		return $query->have_posts();
-	}
-
-	/**
-	 * Export submissions with ajax. Bulk export is a Pro feature.
-	 */
-	public function export_submissions() {
-		if ( ! Pro::is_pro_active() ) {
-			wp_die( esc_html( __( 'Exporting submissions requires Otter Pro.', 'otter-blocks' ) ) );
-		}
-
-		$nonce = isset( $_POST['_nonce'] ) ? sanitize_text_field( $_POST['_nonce'] ) : ''; // phpcs:ignore WordPress.Security.NonceVerification.Recommended
-		if ( ! wp_verify_nonce( $nonce, 'otter_form_export_submissions' ) ) {
-			wp_die( esc_html( __( 'Invalid nonce.', 'otter-blocks' ) ) );
-		}
-
-		if ( ! current_user_can( 'manage_options' ) ) {
-			wp_die( esc_html( __( 'You are not allowed to export submissions.', 'otter-blocks' ) ) );
-		}
-
-		// Export submissions.
-		require_once ABSPATH . 'wp-admin/includes/export.php';
-		ob_start();
-		export_wp( array( 'content' => self::FORM_RECORD_TYPE ) );
-		$export = ob_get_clean();
-
-		echo ent2ncr( $export ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
-		wp_die();
 	}
 
 	/**
